requirements.txt
matplotlib==3.6.3
pandas==2.1.4+dfsg
python_stdnum==1.20
termcolor==2.4.0
yfinance==0.2.37
